% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_collection.R
\docType{methods}
\name{as.collection}
\alias{as.collection}
\alias{as.collection,list-method}
\title{Turn a list of Expressions, Terms or Documents into a Collection}
\usage{
as.collection(list)

\S4method{as.collection}{list}(list)
}
\arguments{
\item{x}{list of which elements are objects that belong to the Expression, Term or Document classes of the sfutils package}
}
\value{
Object of class \code{\link[sfutils]{Collection}}
}
\description{
Turn a list of Expressions, Terms or Documents into a Collection
}
\examples{
\dontrun{
# Get data
data("company_descriptions")

# Put text in a list
txt <- lapply(company_descriptions, function(x) x$desc)

# Fingerprint documents
txt_fp <- do_fingerprint_document(txt)

# To list
lst <- as.list(txt_fp)

# To collection
colcol <- as.collection(lst)
}
}
