% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_collection.R
\docType{methods}
\name{as.list,Collection-method}
\alias{as.list,Collection-method}
\title{Turn a Collection into a list}
\usage{
\S4method{as.list}{Collection}(x, ...)
}
\arguments{
\item{x}{object of class Collection}
}
\value{
List of Collection entries
}
\description{
Turn a Collection into a list
}
