% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_collection.R
\docType{methods}
\name{[[,Collection-method}
\alias{[[,Collection-method}
\title{Indexing of a Collection object}
\usage{
\S4method{[[}{Collection}(x, i, j, drop)
}
\arguments{
\item{object}{Collection object}
}
\description{
Indexing of a Collection object
}
