% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_collection.R
\docType{methods}
\name{length,Collection-method}
\alias{length,Collection-method}
\title{Length of a Collection object}
\usage{
\S4method{length}{Collection}(x)
}
\arguments{
\item{object}{Collection object}
}
\description{
Length of a Collection object
}
