% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_collection.R
\docType{methods}
\name{plot,Collection-method}
\alias{plot,Collection-method}
\title{Plot method for Collection class}
\usage{
\S4method{plot}{Collection}(x, y)
}
\arguments{
\item{object}{object to plot}
}
\description{
Plot method for Collection class
}
