% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_collection.R
\docType{methods}
\name{show,Collection-method}
\alias{show,Collection-method}
\title{Show method for Collection class}
\usage{
\S4method{show}{Collection}(object)
}
\arguments{
\item{object}{object to print}
}
\description{
Show method for Collection class
}
