% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_collection.R
\docType{methods}
\name{[,Collection,ANY,ANY,ANY-method}
\alias{[,Collection,ANY,ANY,ANY-method}
\title{Indexing of a Collection object}
\usage{
\S4method{[}{Collection,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{object}{Collection object}
}
\description{
Indexing of a Collection object
}
