% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{do_create_filter}
\alias{do_create_filter}
\title{Create a category filter}
\usage{
do_create_filter(name, positive, negative = c())
}
\arguments{
\item{name}{name of the filter}

\item{positive}{positive examples. Must consist of a vector with at least two strings.}

\item{negative}{optional negative examples. If you don't want to use these leave them empty.}
}
\value{
an object of class Filter
}
\description{
This function is a wrapper. This should be the only function used by the end user to create a category filter
}
\seealso{
See the \code{\link{Filter-class}} documentation for examples and more information
}
