% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{do_fingerprint_document}
\alias{do_fingerprint_document}
\title{Fingerprint one or multiple texts}
\usage{
do_fingerprint_document(x)
}
\arguments{
\item{x}{single text or list of texts to be fingerprinted.}
}
\value{
either a single instance of Document class (if single text is passed) or an instance of Collection class containing multiple Documents (if a list of texts is passed)
}
\description{
This function is a wrapper. This should be the only function used by the end user to fingerprint texts ('documents')
}
\seealso{
See the \code{\link{Document-class}} documentation for examples and more information
}
