% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{do_fingerprint_expression}
\alias{do_fingerprint_expression}
\title{Create an expression}
\usage{
do_fingerprint_expression(x)
}
\arguments{
\item{x}{an expression or list of expressions to be fingerprinted}
}
\value{
Either an object of class Expression (if user passes character) or an object of class Collection containing multiple expressions (if user passes list)
}
\description{
This function is a wrapper. This should be the only function used by the end user to create an expression
}
\examples{
\dontrun{
# Get a single expression from API. Use subtract param
 body <- list(
  "sub" = list(
    "term" = "jaguar",
    "positions" = list(2,3,4,5,6)
  )
)

# Get multiple expressions from API.
body <- list(
  list("sub" = list(
   "term" = "jaguar",
   "positions" = list(2,3,4,5,6)
  )),
  list("text" = "moodie moodle education whatever I am cool",
    "sub" = list(
      "term" = "online",
      "positions" = list(2,3,9,54)
    )),
  list("text" = "example text I am")
)
}

}
\seealso{
See the \code{\link{Expression-class}} documentation for examples and more information
}
