% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{do_fingerprint_term}
\alias{do_fingerprint_term}
\title{Fingerprint a term}
\usage{
do_fingerprint_term(x)
}
\arguments{
\item{x}{a term to fingerprint}
}
\value{
an object of class Term
}
\description{
This function is a wrapper. This should be the only function used by the end user to fingerprint terms
}
\seealso{
See the \code{\link{Term-class}} documentation for examples and more information
}
