% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_collection.R
\docType{methods}
\name{do_search_collection}
\alias{do_search_collection}
\alias{do_search_collection,Collection-method}
\title{Search a collection for one or multiple unique ids}
\usage{
do_search_collection(obj, uuid)

\S4method{do_search_collection}{Collection}(obj, uuid)
}
\arguments{
\item{obj}{Collection object to search in}

\item{uuid}{one or multiple unique ids to search for}
}
\value{
Object of class \code{\link[sfutils]{Collection}} containing the objects that were searched for
}
\description{
Search a collection for one or multiple unique ids
}
\examples{
\dontrun{
# Get data
data("company_descriptions")

# Put text in a list
txt <- lapply(company_descriptions, function(x) x$desc)

# Fingerprint documents
txt_fp <- do_fingerprint_document(unname(txt))

# Search for
}
}
