% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_document.R
\docType{methods}
\name{do_slice_document}
\alias{do_slice_document}
\alias{do_slice_document,Document-method}
\title{Slice a text}
\usage{
do_slice_document(object, ...)

\S4method{do_slice_document}{Document}(object)
}
\arguments{
\item{object}{An object of class Document}
}
\value{
list of length n, where n is the number of slices.
}
\description{
Returns an ordered list of text objects (ordered according to where the text slice appears in the input text). A Text object consists of a text slice (defined as a slice by the Retina) and a Fingerprint object corresponding to the text slice.
}
