% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_document.R
\docType{methods}
\name{do_tokenize}
\alias{do_tokenize}
\alias{do_tokenize,Document-method}
\title{Tokenize a text}
\usage{
do_tokenize(object, ...)

\S4method{do_tokenize}{Document}(object)
}
\arguments{
\item{object}{An object of class Document}
}
\value{
vector of length n. Here, n is equal to the number of sentences in the text. Each sentence is tokenized using commas.
}
\description{
Tokenize a text using the Cortical API
}
