% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_core.R
\docType{class}
\name{Fingerprint-class}
\alias{Fingerprint-class}
\alias{.Fingerprint}
\title{Fingerprint class}
\description{
This class forms the basis for the core cortical API classes (Document, Term, Expression, Filter). It should not be called by the user directly.
}
\section{Slots}{

\describe{
\item{\code{uuid}}{unique id for the fingerprint. Can be passed, else generated.}

\item{\code{type}}{type of fingerprint (e.g. document if text, filter if filter)}
}}

\seealso{
See the \href{http://documentation.cortical.io/intro.html}{Cortical documentation} for more information about semantic fingerprinting
}
