% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fingerprint.R
\docType{methods}
\name{plot,Fingerprint-method}
\alias{plot,Fingerprint-method}
\title{Plot method for Fingerprint class}
\usage{
\S4method{plot}{Fingerprint}(x, y, ...)
}
\arguments{
\item{x}{object of class Fingerprint to plot}

\item{y}{optional. object of class Fingerprint to plot}

\item{...}{other options. You can pass the results of 'get_context()' or 'get_similar_terms()' to highlight them in the plot.}
}
\description{
Plot method for Fingerprint class
}
