% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fingerprint.R
\docType{methods}
\name{show,Fingerprint-method}
\alias{show,Fingerprint-method}
\title{Print method for Fingerprint class}
\usage{
\S4method{show}{Fingerprint}(object)
}
\arguments{
\item{object}{object to print}
}
\description{
Print method for Fingerprint class
}
