% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_document.R
\docType{methods}
\name{get_keywords}
\alias{get_keywords}
\alias{get_keywords,Document-method}
\title{Retrieve a list of keywords for a text}
\usage{
get_keywords(object)

\S4method{get_keywords}{Document}(object)
}
\arguments{
\item{object}{a Document class object}
}
\value{
list of keywords
}
\description{
This function retrieves 10 keywords for the text
}
\examples{
\dontrun{
# Get data
data("company_descriptions")

# Get one text
txt <- company_descriptions$unilever$desc

# Fingerprint document
txt_fp <- do_fingerprint_document(txt)

# Get keywords
kw <- get_keywords(txt_fp)
}
}
