% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{get_retinas}
\alias{get_retinas}
\title{Retrieve all available retinas from the Cortical server}
\usage{
get_retinas()
}
\value{
Named list with retina name, description, number of terms, number of rows and columns.
}
\description{
Retrieve all available retinas from the Cortical server
}
\seealso{
See the \href{http://www.cortical.io/resources_apidocumentation.html}{Cortical documentation} for more information about the retinas.
}
