% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{get_similar_terms}
\alias{get_similar_terms}
\title{Get similar terms for a fingerprint}
\usage{
get_similar_terms(object, ...)
}
\arguments{
\item{object}{Either: a single Term or Expression class object OR a Collection containing Term or Expression class objects}
}
\value{
list containing different contexts
}
\description{
Retrieve similar terms for a Term or an Expression object
}
