% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{uuid}
\alias{uuid}
\alias{type}
\alias{fingerprint}
\alias{text}
\alias{sfexpression}
\alias{term}
\alias{df}
\alias{score}
\alias{pos_types}
\alias{name}
\alias{positive}
\alias{negative}
\alias{entries}
\alias{uuid,Fingerprint-method}
\alias{fingerprint,Fingerprint-method}
\alias{type,Fingerprint-method}
\alias{text,Document-method}
\alias{sfexpression,Expression-method}
\alias{term,Term-method}
\alias{df,Term-method}
\alias{score,Term-method}
\alias{pos_types,Term-method}
\alias{name,Filter-method}
\alias{positive,Filter-method}
\alias{negative,Filter-method}
\alias{entries,Collection-method}
\title{Get slots from FinTxtUtuls classes}
\usage{
uuid(object)

type(object)

fingerprint(object)

text(object)

sfexpression(object)

term(object)

df(object)

score(object)

pos_types(object)

name(object)

positive(object)

negative(object)

entries(object)

\S4method{uuid}{Fingerprint}(object)

\S4method{fingerprint}{Fingerprint}(object)

\S4method{type}{Fingerprint}(object)

\S4method{text}{Document}(object)

\S4method{sfexpression}{Expression}(object)

\S4method{term}{Term}(object)

\S4method{df}{Term}(object)

\S4method{score}{Term}(object)

\S4method{pos_types}{Term}(object)

\S4method{name}{Filter}(object)

\S4method{positive}{Filter}(object)

\S4method{negative}{Filter}(object)

\S4method{entries}{Collection}(object)
}
\arguments{
\item{object}{a class object containing the slot}
}
\description{
These functions retrieve slots from FinTxtUtuls classes (Fingerprint, Term, Document, Expression, Collection).
}
