% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ancillary.R
\name{Intensity}
\alias{Intensity}
\title{Constructor for Intensity class}
\usage{
Intensity(fp_list)
}
\arguments{
\item{fp_list}{list of objects with a fingerprint}
}
\description{
Constructor for Intensity class
}
\author{
Jasper Ginn
}
