% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{company_descriptions}
\alias{company_descriptions}
\title{Small sample of business descriptions of six companies}
\format{A list, each entry of which is a list (one per company)
\describe{
\item{desc}{business description}
\item{name}{name of the company}
\item{ticker}{ticker of the company}
\item{url}{url to the company page on Thompson Reuters website}
}}
\source{
\url{https://www.reuters.com/markets/stocks}
}
\usage{
company_descriptions
}
\description{
Name, ticker, url and business description of 6 companies. Taken from Thompson Reuters.
}
\keyword{datasets}
