% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAP500}
\alias{SAP500}
\title{Business descriptions for S&P500 companies}
\format{A list, each entry of which is a list (one per company)
\describe{
\item{company}{name of the company}
\item{url}{url to the company page on Thompson Reuters website}
\item{desc}{business description}
\item{ticker}{ticker of the company}
}}
\source{
\url{https://www.reuters.com/finance/markets/index/.SPX?sortBy=last&sortDir=DESC&pn=}
}
\usage{
SAP500
}
\description{
Name, ticker, url and business description of 500 companies. Taken from Thompson Reuters. (downloaded on 27/05/2018)
}
\keyword{datasets}
