% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_core.R
\docType{class}
\name{Term-class}
\alias{Term-class}
\alias{.Term}
\alias{Term}
\title{Term class}
\usage{
Term(term, ...)
}
\arguments{
\item{term}{term to be fingerprinted}

\item{...}{other options to be passed (uuid)}
}
\description{
The Term class is one of the four core classes in the sfutils package. A Term is a single word.
}
\details{
(From \href{Cortical documentation}{http://documentation.cortical.io/working_with_terms.html}) The basic building blocks for performing semantic computations are the representations for single terms. Each Retina contains semantic representations (fingerprints) for a large number of terms, and this page describes how to retrieve these from the API. Furthermore we describe how to query the Retina for semantically similar terms, and retrieve a list of contexts in which a given term can occur.
}
\section{Slots}{

\describe{
\item{\code{term}}{term to be fingerprinted}

\item{\code{df}}{the df value of the term}

\item{\code{score}}{the score of this term}

\item{\code{pos_types}}{the position type of this term}

\item{\code{fingerprint}}{numeric vector of the fingerprint}
}}

\examples{
\dontrun{
# Fingerprint a term
trm_fp <- do_fingerprint_term("Finance")
}
}
\seealso{
See the \href{http://documentation.cortical.io/working_with_terms.html}{Cortical documentation} for more information about semantic fingerprinting and terms
}
