% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_collection.R,
%   R/methods_expression.R, R/methods_term.R
\docType{methods}
\name{get_context}
\alias{get_context}
\alias{get_context,Collection-method}
\alias{get_context,Expression-method}
\alias{get_context,Term-method}
\title{Get contexts for a fingerprint}
\usage{
get_context(object, ...)

\S4method{get_context}{Collection}(object)

\S4method{get_context}{Expression}(object, startIndex = 0L,
  maxResults = 10L, sparsity = 1)

\S4method{get_context}{Term}(object)
}
\arguments{
\item{object}{Either: a single Term or Expression class object OR a Collection containing Term or Expression class objects}

\item{startIndex}{ContextId from which to start}

\item{maxResults}{Maximum number of contexts to be returned}

\item{sparsity}{What sparsity should be applied? e.g. 0.02 returns top 2% of fingerprints}
}
\value{
list containing different contexts
}
\description{
Retrieve context for a Term or an Expression context
}
\examples{
\dontrun{

###### FOR TERMS

# Fingerprint a term
trm_fp <- do_fingerprint_term("finance")

# Get contexts
trm_fp_cont <- get_context(trm_fp)

###### FOR EXPRESSIONS

# Get data
data("company_descriptions")

# Put text in a list to be fingerprinted as expression
txt <- lapply(company_descriptions, function(x) list("text" = x$desc))
# Must be an unnamed list
txt <- unname(txt)

# Fingerprint as expression
txt_fp <- do_fingerprint_expression(txt)

# Get contexts
txt_fp_cont <- get_context(txt_fp[[1]])

###### FOR COLLECTIONS

# Get data
data("company_descriptions")

# Put text in a list
txt <- lapply(company_descriptions, function(x) list("text" = x$desc))
# Must be an unnamed list
txt <- unname(txt)

# Fingerprint as expression
txt_fp <- do_fingerprint_expression(txt)

# Get contexts
# This takes quite a while
txt_fp_cont <- get_context(txt_fp, type = "Expression")

}
}
\seealso{
\href{http://documentation.cortical.io/the_power_of_expressions.html}{This} page for more information about contexts and expressions

\href{http://documentation.cortical.io/working_with_terms.html}{This} page for more information about contexts and terms
}
