% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_expression.R,
%   R/methods_term.R
\docType{methods}
\name{get_similar_terms}
\alias{get_similar_terms}
\alias{get_similar_terms,Expression-method}
\alias{get_similar_terms,Term-method}
\title{Get similar terms for a fingerprint}
\usage{
get_similar_terms(object, ...)

\S4method{get_similar_terms}{Expression}(object, contextId = NULL,
  posType = NULL, startIndex = 0L)

\S4method{get_similar_terms}{Term}(object, contextId = NULL, posType = NULL,
  startIndex = 0L)
}
\arguments{
\item{object}{Either: a single Term or Expression class object OR a Collection containing Term or Expression class objects}
}
\value{
list containing different contexts
}
\description{
Retrieve similar terms for a Term or an Expression object
}
\seealso{
\href{http://documentation.cortical.io/the_power_of_expressions.html}{This} page for more information about contexts and expressions

\href{http://documentation.cortical.io/working_with_terms.html}{This} page for more information about contexts and terms
}
